requirements_test.txt
aionefit==0.14
pytest
pytest-cov
pytest-homeassistant-custom-component==0.5.12
